requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
Django==5.1.1
django-allauth==64.2.0
django-cors-headers==4.4.0
django-htmx==1.19.0
django-stubs==5.0.4
django-stubs-ext==5.0.4
django-tailwind==3.8.0
django-unfold==0.37.0
django-widget-tweaks==1.5.0
fontawesomefree==6.6.0
fonttools==4.53.1
idna==3.8
joblib==1.4.2
kaleido==0.2.1
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.1
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.24.1
pycparser==2.22
PyJWT==2.9.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tenacity==9.0.0
threadpoolctl==3.5.0
tomli==2.0.1
types-PyYAML==6.0.12.20240808
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
